# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ScholarForge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupBody {
  name: string;
  email: string;
  password: string;
  /** @nullable */
  institution?: string | null;
  /** @nullable */
  researchInterests?: string | null;
}

export interface SigninBody {
  email: string;
  password: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  institution?: string | null;
  /** @nullable */
  researchInterests?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  image?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UpdateUserBody {
  name?: string;
  /** @nullable */
  institution?: string | null;
  /** @nullable */
  researchInterests?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  image?: string | null;
}

export interface UpdateUserRoleBody {
  role: string;
}

export interface ChangePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export interface Project {
  id: string;
  title: string;
  description: string;
  /** @nullable */
  abstract?: string | null;
  keywords: string[];
  status: string;
  visibility: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  createdAt: string;
  updatedAt: string;
  memberCount: number;
  /** @nullable */
  currentUserRole?: string | null;
}

export interface ProjectSummary {
  id: string;
  title: string;
  description: string;
  keywords: string[];
  status: string;
  visibility: string;
  memberCount: number;
  taskCount: number;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  currentUserRole?: string | null;
}

export interface CreateProjectBody {
  title: string;
  description: string;
  /** @nullable */
  abstract?: string | null;
  keywords: string[];
  status: string;
  visibility: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}

export interface UpdateProjectBody {
  title?: string;
  description?: string;
  /** @nullable */
  abstract?: string | null;
  keywords?: string[];
  status?: string;
  visibility?: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}

export interface ProjectMember {
  id: string;
  userId: string;
  projectId: string;
  role: string;
  joinedAt: string;
  user: User;
}

export interface UpdateMemberBody {
  role: string;
}

export interface Invitation {
  id: string;
  email: string;
  projectId: string;
  invitedById: string;
  status: string;
  createdAt: string;
  project: ProjectSummary;
  invitedBy: User;
}

export interface CreateInvitationBody {
  email: string;
}

export interface Task {
  id: string;
  projectId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  /** @nullable */
  assignedToId?: string | null;
  /** @nullable */
  dueDate?: string | null;
  assignedTo?: User | null;
}

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  status: string;
  /** @nullable */
  assignedToId?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  status?: string;
  /** @nullable */
  assignedToId?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface Milestone {
  id: string;
  projectId: string;
  title: string;
  dueDate: string;
  completed: boolean;
}

export interface CreateMilestoneBody {
  title: string;
  dueDate: string;
}

export interface UpdateMilestoneBody {
  title?: string;
  dueDate?: string;
  completed?: boolean;
}

export interface ProjectFile {
  id: string;
  projectId: string;
  filename: string;
  fileUrl: string;
  uploadedById: string;
  uploadedAt: string;
  uploadedBy: User;
}

export interface UploadFileBody {
  filename: string;
  fileUrl: string;
}

export interface ChatMessage {
  id: string;
  projectId: string;
  userId: string;
  message: string;
  createdAt: string;
  user: User;
}

export interface SendMessageBody {
  message: string;
}

export interface ActivityLog {
  id: string;
  projectId: string;
  userId: string;
  action: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
  user: User;
}

export interface PlatformStats {
  totalUsers: number;
  totalProjects: number;
  activeProjects: number;
  completedProjects: number;
  draftProjects: number;
  publicProjects: number;
  totalMessages: number;
  totalFiles: number;
  recentSignups: number;
}

export interface DashboardSummary {
  myProjectCount: number;
  collaboratingCount: number;
  pendingInvitations: number;
  recentProjects: ProjectSummary[];
  recentActivity: ActivityLog[];
}

export interface ProjectStats {
  totalTasks: number;
  completedTasks: number;
  inProgressTasks: number;
  todoTasks: number;
  totalMilestones: number;
  completedMilestones: number;
  memberCount: number;
  fileCount: number;
  messageCount: number;
}

export type ListUsersParams = {
  search?: string;
  role?: string;
};

export type ListProjectsParams = {
  visibility?: string;
  status?: string;
  search?: string;
  myProjects?: string;
};

export type ListTasksParams = {
  status?: string;
};

export type ListMessagesParams = {
  limit?: number;
  before?: string;
};

export type ListActivityParams = {
  limit?: number;
};