# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScholarForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signin:
    post:
      operationId: signin
      tags: [auth]
      summary: Sign in with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SigninBody"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Users
...
  /projects/{projectId}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload a file to a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadFileBody"
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
  /projects/{projectId}/files/{fileId}:
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: path
          name: fileId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: File deleted

  # Chat Messages
...
  /projects/{projectId}/activity:
    get:
      operationId: listActivity
      tags: [activity]
      summary: Get activity log for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Activity log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"

  # Analytics / Dashboard
...
  /analytics/projects/{projectId}/stats:
    get:
      operationId: getProjectStats
      tags: [analytics]
      summary: Get stats for a specific project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectStats"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SignupBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        institution:
          type: ["string", "null"]
        researchInterests:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
    SigninBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # User
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        institution:
          type: ["string", "null"]
        researchInterests:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
        - updatedAt
...
    ProjectFile:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        filename:
          type: string
        fileUrl:
          type: string
        uploadedById:
          type: string
        uploadedAt:
          type: string
        uploadedBy:
          $ref: "#/components/schemas/User"
      required:
        - id
        - projectId
        - filename
        - fileUrl
        - uploadedById
        - uploadedAt
        - uploadedBy
    UploadFileBody:
      type: object
      properties:
        filename:
          type: string
        fileUrl:
          type: string
      required:
        - filename
        - fileUrl

    # Chat
...
    ActivityLog:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        userId:
          type: string
        action:
          type: string
        details:
          type: ["string", "null"]
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - projectId
        - userId
        - action
        - createdAt
        - user

    # Analytics
...
    ProjectStats:
      type: object
      properties:
        totalTasks:
          type: integer
        completedTasks:
          type: integer
        inProgressTasks:
          type: integer
        todoTasks:
          type: integer
        totalMilestones:
          type: integer
        completedMilestones:
          type: integer
        memberCount:
          type: integer
        fileCount:
          type: integer
        messageCount:
          type: integer
      required:
        - totalTasks
        - completedTasks
        - inProgressTasks
        - todoTasks
        - totalMilestones
        - completedMilestones
        - memberCount
        - fileCount
        - messageCount